✅ Static domain class added for ORM ecs entities ✅ ECS integration in BattleField Generation ✅  Added Resource folder with temp sprite files

+++ Assets/Project/ECS/Components/BattleFieldCells/BattleFieldCellPosition.cs
@@ -7,7 +7,7 @@ namespace ECS.Components{
         public int grid_x;
         public int grid_y;
         
-        public int global_x;
-        public int global_y; 
+        public float global_x;
+        public float global_y; 
     }
 }
